+++ src/Fable.Core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* [TS/JS] Add `Pojo` attribute (by @alfonsogarciacaro)
+
 ## 4.3.0 - 2024-01-25
 
 ### Added